requirements.txt
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tiktoken>=0.5.0
cryptography>=41.0.0